@@ -13,7 +13,7 @@ class AnimatedLogo extends AnimatedWidget {
   static final _opacityTween = Tween<double>(begin: 0.1, end: 1);
   static final _sizeTween = Tween<double>(begin: 0, end: 300);
 
-  AnimatedLogo({Key key, Animation<double> animation})
+  AnimatedLogo({Key? key, required Animation<double> animation})
       : super(key: key, listenable: animation);
 
   Widget build(BuildContext context) {
@@ -37,8 +37,8 @@ class LogoApp extends StatefulWidget {
 }
 
 class _LogoAppState extends State<LogoApp> with SingleTickerProviderStateMixin {
-  Animation<double> animation;
-  AnimationController controller;
+  late Animation<double> animation;
+  late AnimationController controller;
 
   @override
   void initState() {
@@ -107,7 +107,7 @@ class UsedInTutorialTextOnly extends _LogoAppState {
     // #docregion IntTween-curve
     AnimationController controller = AnimationController(
         duration: const Duration(milliseconds: 500), vsync: this);
-    final Animation curve =
+    final Animation<double> curve =
         CurvedAnimation(parent: controller, curve: Curves.easeOut);
     Animation<int> alpha = IntTween(begin: 0, end: 255).animate(curve);
     // #enddocregion IntTween-curve
